+++ high_five.rb
@@ -1,0 +1,7 @@
+module HighFiveable
+
+def offer_high_five
+    "High five!"
+end
+
+end
